Initial version of the react integration
- We check in the JSX file of the tutorial directly into e-mission

- We include ngReact in the dependencies
- We include `<script` tags for the react libraries and ngReact itself
- We include babel in the npm dependencies
- We add a command `compile-react` to run babel to compile the JSX
- We add a README instruction on how to run it and generate the js file
- We change the app to depend on react and to create a directive with the
  `TutorialComponent`
- We add a new file `survey.html` which basically includes only the `<tutorial`
  tag in the main body
- We read the file into the scope taking i18n into account, similar to the
  other files in the intro
- We add a new slide for the new `survey.html`, using the scope variable, of course

With this change, we can see tutorial during the onboarding
https://github.com/e-mission/e-mission-docs/issues/643#issuecomment-877983937

This is consistent with my action item in
https://github.com/e-mission/e-mission-docs/issues/643#issuecomment-878505316

Note that since `package.cordovabuild.json` currently does not include any
commands, this is only supported in the devapp (serve) option. using a full
native build will not allow react components to be added.

This is fine since this is a temporary measure, which will be addressed when we
break this dependency.